fixup! refactor(ivy): use `ClassDeclaration` in more `ReflectionHost` methods
