+++ packages/compiler-cli/ngcc/test/host/esm2015_host_spec.ts
@@ -8,11 +8,11 @@
 
 import * as ts from 'typescript';
 
-import {ClassMemberKind, Import} from '../../../src/ngtsc/reflection';
+import {ClassMemberKind, Import, isNamedClassDeclaration, isNamedFunctionDeclaration, isNamedVariableDeclaration} from '../../../src/ngtsc/reflection';
 import {Esm2015ReflectionHost} from '../../src/host/esm2015_host';
 import {getDeclaration, makeTestBundleProgram, makeTestProgram} from '../helpers/utils';
 
-import {expectTypeValueReferencesForParameters, isNamedClassDeclaration, isNamedFunctionDeclaration, isNamedVariableDeclaration} from './util';
+import {expectTypeValueReferencesForParameters} from './util';
 
 const SOME_DIRECTIVE_FILE = {
   name: '/some_directive.js',
