+++ packages/compiler-cli/src/ngtsc/annotations/test/BUILD.bazel
@@ -21,7 +21,6 @@ ts_library(
         "//packages/compiler-cli/src/ngtsc/scope",
         "//packages/compiler-cli/src/ngtsc/testing",
         "//packages/compiler-cli/src/ngtsc/translator",
-        "//packages/compiler-cli/src/ngtsc/util",
         "@npm//typescript",
     ],
 )
