+++ packages/compiler-cli/src/ngtsc/annotations/test/directive_spec.ts
@@ -10,10 +10,9 @@ import * as ts from 'typescript';
 
 import {NOOP_DEFAULT_IMPORT_RECORDER, ReferenceEmitter} from '../../imports';
 import {PartialEvaluator} from '../../partial_evaluator';
-import {ClassDeclaration, TypeScriptReflectionHost} from '../../reflection';
+import {ClassDeclaration, TypeScriptReflectionHost, isNamedClassDeclaration} from '../../reflection';
 import {LocalModuleScopeRegistry, MetadataDtsModuleScopeResolver} from '../../scope';
 import {getDeclaration, makeProgram} from '../../testing/in_memory_typescript';
-import {isNamedClassDeclaration} from '../../util/src/typescript';
 import {DirectiveDecoratorHandler} from '../src/directive';
