+++ packages/compiler-cli/src/ngtsc/reflection/index.ts
@@ -9,3 +9,4 @@
 export * from './src/host';
 export {typeNodeToValueExpr} from './src/type_to_value';
 export {TypeScriptReflectionHost, filterToMembersWithDecorator, reflectIdentifierOfDeclaration, reflectNameOfDeclaration, reflectObjectLiteral, reflectTypeEntityToDeclaration} from './src/typescript';
+export {isNamedClassDeclaration, isNamedFunctionDeclaration, isNamedVariableDeclaration} from './src/util';
