+++ packages/compiler-cli/src/ngtsc/reflection/test/BUILD.bazel
@@ -12,7 +12,6 @@ ts_library(
         "//packages:types",
         "//packages/compiler-cli/src/ngtsc/reflection",
         "//packages/compiler-cli/src/ngtsc/testing",
-        "//packages/compiler-cli/src/ngtsc/util",
         "@npm//typescript",
     ],
 )
